댓글달기

+++ src/component/funding/FundingComment.js
@@ -81,6 +81,13 @@ function FundingComment() {
   //댓글수정
   const editComment = async (commentId, updatedContents) => {
     try {
+      const commentWriter = data.find((item) => item.no === commentId)?.name;
+
+      if (commentWriter !== writer) {
+        alert("작성자가 아닙니다.");
+        return;
+      }
+
       await axios.put(
         `${bootPath.bootpath}/funding/comment/update?no=${commentId}`,
         {
@@ -101,6 +108,13 @@ function FundingComment() {
   };
 
   const editClick = (commentId) => {
+    const commentWriter = data.find((item) => item.no === commentId)?.name;
+
+    if (commentWriter !== writer) {
+      alert("작성자가 아닙니다.");
+      return;
+    }
+
     setEditingCommentId(commentId);
   };
 
@@ -120,6 +134,13 @@ function FundingComment() {
   //댓글삭제
   const deleteComment = async (commentId) => {
     try {
+      const commentWriter = data.find((item) => item.no === commentId)?.name;
+
+      if (commentWriter !== writer) {
+        alert("작성자가 아닙니다.");
+        return;
+      }
+
       await axios.delete(
         `${bootPath.bootpath}/funding/comment/delete?no=${commentId}`
       );
